fix: namespacing issues from file rename
