+++ source/RewardsService/rewardsservice/handlers/user_rewards_handler.py
@@ -1,0 +1,5 @@
+from .rewards_base_handler import RewardsBaseHandler
+
+
+class GetUserRewardsHandler(RewardsBaseHandler):
+    pass
