+++ source/RewardsService/rewardsservice/url_patterns.py
@@ -1,5 +1,6 @@
-from handlers.rewards_handler import GetRewardTiersHandler
-from handlers.order_data_handler import UserTierHandler
+from handlers.reward_tiers_handler import GetRewardTiersHandler
+from handlers.user_tier_handler import UserTierHandler
+from handlers.user_rewards_handler import GetUserRewardsHandler
 
 url_patterns = [
     (r"/", GetRewardTiersHandler),
